requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gTTS==2.5.1
transformers==4.41.2
torch>=2.0.0
PyPDF2==3.0.1
Pillow==10.3.0
sentencepiece
protobuf
# how to use ---pip install -r requirements.txt

